pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "mcp>=1.8.1",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
